Scores are working
Added code to find fields in json[responses] but not in schema[embedded_data]; added code to determine fields not account for so that these can be attached to embedded_data in schema and embedded_data_names.

Scores, and other qualtrics variables not present schema, are treated as embedded data which itself is treated as questions.